refactor: リアーキテクチャ/とりあえずディレクトリ分けた
