+++ src/adapters.rs
@@ -1,0 +1,3 @@
+pub mod cli;
+pub mod config;
+pub mod workspace;
