+++ src/adapters/cli.rs
