+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod adapters;
+mod entities;
+pub mod usecases;
